JS: enforce camel case
We seem to be bad at sticking to this rule, so I think it's worth
enforcing it.

I think part of why we haven't done this before is that we were afraid
of triggering this on code interacting with JSON fields from API
responses, which generally *are* underscored. From my local testing,
I actually don't think that's a concern: the rule seems to trigger on
assignment/creation (`var foo_bar`, `{ name_blah: true}`, etc.), but not
access (e.g. `thing.foo_bar()` and `thing["attr_name"]` don't trigger
this).

Thoughts, @codeclimate/review?