Turn off rubocop's FormatStringToken check
This appeared when we recently upgraded RuboCop: I don't think it's a
style we should enforce.

The named tokens in format strings can be useful in very complicated
cases, but those cases are not the norm, and the check's suggested style
makes simple cases *less* readable.

In 0.52.0 it also has bugs leading to many false positives where
changing the code as suggested would result in incorrect behavior:
https://github.com/bbatsov/rubocop/issues/5245.