Use "indented" style for MultilineMethodCallIndentation
The current (default) value for this check wants chained assignments to
look like:

```
foo =
  Foo.
  bar.
  baz
```

That is not a style I think anybody on our team habitually writes. What
I think we tend to do (and which matches the `indented` style) is:

```
foo = Foo.
  bar.
  baz
```

Since AFAIK this is already the most common de-facto style we use, we
should probably set that in the styleguide.